Split up e2e test into separate suites
